Update project structure. Split files
