+++ app/models/currency_converter.rb
@@ -1,35 +1,4 @@
-class OpenExchangeRatesClient
-  include ActiveModel::Validations
-  include ActiveModel::Attributes
-
-  attribute :app_id, :string
-
-  validates :app_id, presence: true
-
-  def base_url
-    URI::HTTPS.build(
-      host: 'openexchangerates.org',
-      path: '/api'
-    )
-  end
-
-  def fetch_currencies
-    currencies_url = base_url.tap do |url|
-      url.path << '/currencies.json'
-    end
-    Net::HTTP.get(currencies_url)
-  end
-
-  # TODO: check having app_id
-  def fetch_historical_for(date:)
-    validate!
-    historical_url = base_url.tap do |url|
-      url.path << "/historical/#{date}.json"
-      url.query = "app_id=#{app_id}"
-    end
-    Net::HTTP.get(historical_url)
-  end
-end
+require_relative './open_exchange_rates_client'
 
 class CurrencyConverter
   include ActiveModel::Model
