+++ boot.rb
@@ -1,0 +1,8 @@
+require 'net/http'
+require 'json'
+require 'date'
+require 'active_model'
+require 'active_support/core_ext'
+
+require_relative './config'
+require_relative './app/models/currency_converter'
