+++ config.rb
@@ -1,10 +1,3 @@
-require 'net/http'
-require 'json'
-require 'date'
-
-require 'active_model'
-require 'active_support/core_ext'
-
 module Config
   autoload :Secrets, './config/secrets'
 end
